+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [UNPUBLISHED]
 
 - transitioned from htmlPreview to webview
+- rerender on configuration change
 
 # 0.10.2 / 2018-12-10